Sr25519 tests
