+++ fearless-utils/src/test/java/jp/co/soramitsu/fearless_utils/bip39/Bip39Test.kt
@@ -1,34 +1,20 @@
 package jp.co.soramitsu.fearless_utils.bip39
 
-import com.google.gson.Gson
-import jp.co.soramitsu.fearless_utils.common.getResourceReader
 import jp.co.soramitsu.fearless_utils.encrypt.EncryptionType
-import jp.co.soramitsu.fearless_utils.encrypt.keypair.substrate.SubstrateKeypairFactory
 import jp.co.soramitsu.fearless_utils.extensions.fromHex
 import jp.co.soramitsu.fearless_utils.extensions.toHexString
-import jp.co.soramitsu.fearless_utils.junction.SubstrateJunctionDecoder
 import org.junit.Assert.assertEquals
-import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.junit.MockitoJUnitRunner
 import org.spongycastle.util.encoders.Hex
 
 @RunWith(MockitoJUnitRunner::class)
-class Bip39Test {
-
-    private lateinit var bip39: Bip39
+class LocalBip39Test : Bip39Test() {
 
     private val expectedEntropyHex = "2a5ecdeb7466f14d3c06d5aa5c6d433d"
     private val expectedMnemonic =
         "clean wait kiss trip humor pledge useless survey prevent toddler express knock"
-
-    private val gson = Gson()
-
-    @Before
-    fun setUp() {
-        bip39 = Bip39()
-    }
 
     @Test
     fun generateMnemonic() {
@@ -52,34 +38,6 @@ class Bip39Test {
         performSpecTests("crypto/ecdsaHDKD.json", EncryptionType.ECDSA)
     }
 
-    private fun performSpecTests(filename: String, encryptionType: EncryptionType) {
-        val testCases = gson.fromJson(
-            getResourceReader(filename),
-            Array<MnemonicTestCase>::class.java
-        )
-
-        testCases.forEach { testCase ->
-            val derivationPathRaw = testCase.path.ifEmpty { null }
-
-            val derivationPath = derivationPathRaw
-                ?.let { SubstrateJunctionDecoder.decode(testCase.path) }
-
-            val actualEntropy = bip39.generateEntropy(testCase.mnemonic)
-            val actualSeed = bip39.generateSeed(actualEntropy, derivationPath?.password)
-
-            val actualKeypair = SubstrateKeypairFactory.generate(
-                seed = actualSeed,
-                junctions = derivationPath?.junctions.orEmpty(),
-                encryptionType = encryptionType
-            )
-
-            assertEquals(
-                "Mnemonic=${testCase.mnemonic}, derivationPath=${testCase.path}",
-                testCase.expectedPublicKey,
-                actualKeypair.publicKey.toHexString(withPrefix = true)
-            )
-        }
-    }
 
     @Test
     fun generateEntropy() {
