+++ fearless-utils/src/testShared/java/jp/co/soramitsu/fearless_utils/bip39/MnemonicTestCase.kt
